chore: re-organize project structure
Refactor the project folder structure to enhance modularity and maintainability. Moved related files into separate directories for modbdels, routes, and controllers. Also, created individual files for authentication routes and product-related routes. The new structure improves code organization and readability, making it easier to manage and scale the application.

This restructuring adheres to best practices, ensuring a clean and organized codebase, making future development and collaboration more efficient.
